Manage topic with a lisp variable
Consider the following scenario: a user sets a topic to some custom variable,
and then makes a pull request.  They make some changes, and then decide to do
another push.  In most cases, it seems reasonable that the user will want to
push to that same PR again.

Thus, we change the behavior of this option such that it is persistent for the
Emacs session through multiple menu invocations, until explicitly changed by the
user.