Don't block when force-pushing a PR
When there is a post buffer to submit, there is data that could be lost if we
delete it too early.  This is not the case when we are force-pushing, so there
is no point of blocking when it that option is enabled.

Use a class object to store pull request information and make it easier to pass
around.