Fix case where incorrect description is used when posting
We were not removing and replacing the description push option before, so
multiple invocations of `agitjo-post-confirm` for the same post buffer would
cause multiple instances of --push-option=description=... to occur in the final
command.  Since the newest descriptions were inserted at the beginning, that
would mean the description from the first `agitjo-post-confirm` invocation would
be used, which is unexpected.